requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
protobuf==3.20.1
SQLAlchemy==1.4.41
Werkzeug==2.2.2
